pyproject.toml
[project]
name = "disaster-surveillance-reporter"
version = "0.1.20260409"
description = "This project digest incident updates from around the world and store the summary"
readme = "README.md"
requires-python = ">=3.13"
license = { file = "LICENSE" }
authors = [
    { name = "eol", email = "nullhack@users.noreply.github.com" }
]
maintainers = [
    { name = "eol", email = "nullhack@users.noreply.github.com" }
]
dependencies = [
    "dotenv>=0.9.9",
    "fire>=0.7.1",
    "google-api-python-client>=2.194.0",
    "gspread>=6.2.1",
    "httpx>=0.28.0",
    "python-dotenv>=1.2.2",
]

[project.urls]
Repository = "https://github.com/nullhack/disaster-surveillance-reporter"
Documentation = "https://github.com/nullhack/disaster-surveillance-reporter/tree/main/docs/api/"

[project.optional-dependencies]
dev = [
    "pdoc>=14.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pytest-html-plus>=0.5",
    "pytest-mock>=3.14.0",
    "ruff>=0.11.5",
    "taskipy>=1.14.1",
    "hypothesis>=6.148.4",
    "cosmic-ray>=8.4.3",
    "pyright>=1.1.407",
]

[tool.setuptools]
packages = ["disaster_surveillance_reporter"]

[tool.ruff.lint]
ignore = []
select = [
  "A", 
  "ANN", 
  "ASYNC", 
  "B", 
  "C4", 
  "C9", 
  "D", 
  "DTZ", 
  "E", 
  "ERA", 
  "F", 
  "FURB", 
  "G", 
  "I", 
  "ICN", 
  "LOG", 
  "N", 
  "NPY", 
  "PD", 
  "PT", 
  "PTH", 
  "R", 
  "RUF", 
  "S", 
  "SIM", 
  "T20", 
  "TD", 
  "W",
]
preview = true
mccabe.max-complexity = 10
pydocstyle.convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "ANN", "D205", "D212", "D415"]

[tool.pytest.ini_options]
minversion = "6.0"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "unit",
    "integration",
    "system",
    "acceptance",
    "regression",
    "smoke",
    "sanity",
    "performance",
    "security",
    "performance",
]
addopts = """
--maxfail=10 \
--color=yes \
--tb=short \
-q \
--html-output=docs/tests \
--json-report=final_report.json \
"""
testpaths = [
    "tests",
    "disaster_surveillance_reporter"
]
python_files = ["*_test.py"]
python_functions = ["test_*"]
render_collapsed = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:"
    ]

[tool.taskipy.tasks]
run = "python -m disaster_surveillance_reporter.disaster_surveillance_reporter"
test-report = """\
pytest \
  --doctest-modules \
  --cov-config=pyproject.toml \
  --cov-report html:docs/coverage \
  --cov-report term:skip-covered \
  --cov=disaster_surveillance_reporter \
  --cov-fail-under=100 \
  --hypothesis-show-statistics \
"""
test = """\
python -c "import subprocess, sys; print('Running Smoke Tests...'); sys.exit(0 if subprocess.run(['pytest', '-m', 'smoke']).returncode in (0,5) else 1)" && \
python -c "import subprocess, sys; print('Running Unit Tests...'); sys.exit(0 if subprocess.run(['pytest', '-m', 'unit', '-m', 'not slow']).returncode in (0,5) else 1)" && \
python -c "import subprocess, sys; print('Running Integration Tests...'); sys.exit(0 if subprocess.run(['pytest', '-m', 'integration', '-m', 'not slow']).returncode in (0,5) else 1)" && \
python -c "print('Running Tests...');" && \
task test-report\
"""
test-fast = """\
python -c "import subprocess, sys; print('Running Smoke Tests...'); sys.exit(0 if subprocess.run(['pytest', '-m', 'smoke']).returncode in (0,5) else 1)" && \
python -c "import subprocess, sys; print('Running Unit Tests...'); sys.exit(0 if subprocess.run(['pytest', '-m', 'unit', '-m', 'not slow']).returncode in (0,5) else 1)" && \
python -c "import subprocess, sys; print('Running Integration Tests...'); sys.exit(0 if subprocess.run(['pytest', '-m', 'integration', '-m', 'not slow']).returncode in (0,5) else 1)" \
"""
test-slow = """\
pytest -m slow \
"""
ruff-check = "ruff check . --fix"
ruff-format = "ruff format ."
lint = "task ruff-check && task ruff-format"
doc-serve = "pdoc ./disaster_surveillance_reporter --host localhost --port 8080"
doc-build = "pdoc ./disaster_surveillance_reporter -o docs/api --search"
doc-publish = """\
task doc-build && \
git checkout gh-pages 2>/dev/null || git checkout -b gh-pages && \
git rm -rf . 2>/dev/null || true && \
cp -r docs/api/* . && \
git add -A && \
git commit -m "Publish API documentation" && \
git push origin gh-pages --force && \
git checkout main"""
mut-report = """
  uv run cosmic-ray new-config mut.toml && \
  uv run cosmic-ray init mut.toml mut.sqlite && \
  uv run cosmic-ray --verbosity=INFO baseline mut.toml && \
  uv run cosmic-ray exec mut.toml mut.sqlite && \
  uv run cr-html mut.sqlite > docs/mut_report.html && \
  rm mut.toml && \
  rm mut.sqlite
"""
static-check = "pyright"
